+++ tests/test_exts_scalars.py
@@ -4,7 +4,7 @@ import uuid
 import pytest
 
 from py_gql.exc import ScalarParsingError, ScalarSerializationError
-from py_gql.exts.scalars import UUID, RegexType
+from py_gql.exts.scalars import UUID, JSONString, RegexType
 from py_gql.lang import parse_value
 
 
@@ -66,3 +66,74 @@ class TestRegexType:
         with pytest.raises(ScalarParsingError) as exc_info:
             t.parse_literal(parse_value("1"))  # type: ignore
         assert str(exc_info.value) == "Invalid literal IntValue"
+
+
+class TestJSONString:
+    @pytest.mark.parametrize(
+        "value,expected",
+        [
+            ('"foo"', "foo"),
+            ("1", 1),
+            ('[1, "foo"]', [1, "foo"]),
+            (
+                '{"name": "foo", "value": 10, "active": false, "modifiers": [null]}',
+                {
+                    "active": False,
+                    "modifiers": [None],
+                    "name": "foo",
+                    "value": 10,
+                },
+            ),
+        ],
+    )
+    def test_parse_ok(self, value, expected):
+        assert expected == JSONString.parse(value)
+
+    @pytest.mark.parametrize(
+        "value,error",
+        [
+            ("", "Expecting value"),
+            ('"foo', "Unterminated"),
+            ("False", "Expecting value"),
+        ],
+    )
+    def test_parse_err(self, value, error):
+        with pytest.raises(ScalarParsingError) as exc_info:
+            JSONString.parse(value)
+        assert error in str(exc_info.value)
+
+    def test_parse_node_invalid_type(self):
+        with pytest.raises(ScalarParsingError):
+            JSONString.parse_literal(parse_value("1"))  # type: ignore
+
+    @pytest.mark.parametrize(
+        "value,expected",
+        [
+            ("foo", '"foo"'),
+            (1, "1"),
+            ([1, "foo"], '[1, "foo"]'),
+            (
+                {
+                    "active": False,
+                    "modifiers": [None],
+                    "name": "foo",
+                    "value": 10,
+                },
+                '{"active": false, "modifiers": [null], "name": "foo", "value": 10}',
+            ),
+        ],
+    )
+    def test_serialize_ok(self, value, expected):
+        assert expected == JSONString.serialize(value)
+
+    @pytest.mark.parametrize(
+        "value,error",
+        [
+            (object(), "not JSON serializable"),
+            (set([1, 2, 3]), "not JSON serializable"),
+        ],
+    )
+    def test_serialize_err(self, value, error):
+        with pytest.raises(ScalarSerializationError) as exc_info:
+            JSONString.serialize(value)
+        assert error in str(exc_info.value)